requirements.txt
numpy==1.26.4
pandas==2.2.2
jenkspy==0.4.1
PyYAML==6.0.1
SQLAlchemy==2.0.34